Rename package avancement to promotion
Also remove unnecessary TODO statement.
