+++ pkg/promotion/pull_request.go
@@ -1,4 +1,4 @@
-package avancement
+package promotion
 
 import (
 	"fmt"
