+++ pkg/promotion/pull_request_test.go
@@ -1,4 +1,4 @@
-package avancement
+package promotion
 
 import (
 	"testing"
